# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Automation SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{projectId}/tasks:
    get:
      operationId: listProjectTasks
      tags: [projects]
      summary: List project tasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectTask"
    post:
      operationId: createProjectTask
      tags: [projects]
      summary: Create project task
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTask"
...
  /analytics/projects:
    get:
      operationId: getProjectAnalytics
      tags: [analytics]
      summary: Project analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Project analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAnalytics"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Project:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        departmentId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        budget:
          type: number
          nullable: true
        progress:
          type: integer
        managerId:
          type: integer
          nullable: true
        managerName:
          type: string
          nullable: true
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - status
        - priority
        - progress
        - taskCount
        - completedTaskCount
        - createdAt
    CreateProjectRequest:
      type: object
      properties:
        departmentId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        budget:
          type: number
          nullable: true
        managerId:
          type: integer
          nullable: true
      required:
        - name
        - status
        - priority
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        budget:
          type: number
          nullable: true
        progress:
          type: integer
        managerId:
          type: integer
          nullable: true
    ProjectTask:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assigneeId:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - createdAt
    CreateProjectTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assigneeId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
      required:
        - title
        - status
        - priority
...
    ProjectAnalytics:
      type: object
      properties:
        totalProjects:
          type: integer
        completedThisMonth:
          type: integer
        overdueProjects:
          type: integer
        avgCompletionRate:
          type: number
        projectsByPriority:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
      required:
        - totalProjects
        - completedThisMonth
        - overdueProjects
        - avgCompletionRate
        - projectsByPriority
...
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count
...